+++ lib/utils/has_feature.dart
